all: move extention to sub directory
This follows a pattern established by vscode-go. Whilst we don't
strictly need a submodule like vscode-go, having the extension itself
allows helper code (Go, CUE etc) to exist in a different directory
structure, with code generate etc into the extension directory as
required.

Also establish a baseline TypeScript-based extension in that
subdirectory as a basis for later changes that will start cuepls etc.

Rather than basing this template on vscode-go, the skeleton was
generated following the instructions at:

    https://code.visualstudio.com/api/get-started/your-first-extension

Specifically, we ran the following command:

    npx --package yo --package generator-code -- yo code

Boilerplate files that we don't need/want for now are placed in
extension/_generated for later reference (but will likely just be
removed).

Tested in VSCode itself to ensure we retain the language syntax
highlighting, and that we can run the "Welcome to CUE" command.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I6b78db4898773da905e990c2752b755fbe816ffc