internal/ci: re-baseline base from cue-lang/cue@82995bbb
This moves us to use the Central Registry (temporary namespace)
definitions for GitHub Actions.

We also copy the current pattern for "authenticating" a step that
requires access to the Central Registry. Not the cleanest of patterns,
but changing that does not belong in this CL.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I00b42fabdbec6dee8608974e86122724dce2fa9d
Reviewed-on: https://review.gerrithub.io/c/cue-lang/vscode-cue/+/1200740
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>