all: gitignore extension directory ahead of stack of changes
Add /extension to the .gitignore file in preparation of the introduction
of that directory later in this stack. Ignore this early so that make
changes in the stack doesn't leave that directory untracked (because in
a later CL we only ignore subdirectories of /extension, which leaves the
directory non-empty).

Also ignore VSCode workspace files - these are, for now, more of a local
thing within this repo.

Also ignore node_modules. We don't ever want to vendor these.

While we are here, establish an early .vscodeignore that should largely
mirror the .gitignore.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I6267702b18b1319f760d5cbffef8d251cee79b70
Reviewed-on: https://review.gerrithub.io/c/cue-lang/vscode-cue/+/1200739
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>