extension: introduce extension configuration
This builds on the site configuration, and for now relies on the common
parts of the site configuration living in the root, and us using the k8s
pattern of instances in subdirectories being evaluated, unified with
packages of the same name to the module root.

However, the pattern we really want to follow he is of a single package
value split across multiple subdirectories, a la the cuelang.org repo.

That requires a solution to https://cuelang.org/issue/2000, more
specifically taking advantage of a config option that allows us to
specify that load mode.

So for now, the k8s style just about works.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I745ba2a61fd464a30f8a27b4e31e8a1d31c7f54a