extension: introduce extension configuration
Establish an extension package with configuration derived from the root
project configuration. For now this is largely a template of the
package.json file, split into two parts:

1. The bulk of the extension configuration in extension.cue
2. The NPM-controlled section of package.json in npm.cue

Separating npm.cue from extension.cue allows for the former to be
"reverse engineered" from package.json, which might be preferable if
package.json has been directly updated by npm commands, e.g.
installaling a new package. To support this, a new CUE workflow command
writebackPackageJSON is provided.

Otherwise, the extension package serves as the basis for code generating
package.json. The CUE workflow command genPackageJSON is provided for
that purpose.

We run genPackageJSON as pre-step before 'vsce package', in order that
package.json is current. As this CL shows, there is no diff in
package.json itself which confirms this step is a no-op in the eyes of
vsce.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I745ba2a61fd464a30f8a27b4e31e8a1d31c7f54a