all: add site configuration and use from CI
This is the beginning of a configuration that will be used to drive
various parts of the project. For now we import the site configuration
as part of the CI configuration, but it could/should really be that the
CI configuration is just a part of the site configuration. That can wait
until later.

Whilst here, drop the use of dummyDispatch for the trybot workflows. We
don't need that for now and it makes "testing" the GitHub workflows via
'cue export' impossible, because of incomplete values.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Ic22934d6939ae3e8bea482baeefaac5706143a78