all: add project configuration and use from CI
This is the beginning of a configuration that will be used to drive
various parts of the project. For now we import the project
configuration as part of the CI configuration, but it could/should
really be that the CI configuration is just a part of the project
configuration. That can wait until later.

Note also that the configuration that exists in project.cue is part of a
package named 'extension'. This is short-lived. In a later CL we move
the extension to a subdirectory, where that package name is retained,
and the root of the project is given the 'project' package name.

Whilst here, drop the use of dummyDispatch for the trybot workflows. We
don't need that for now and it makes "testing" the GitHub workflows via
'cue export' impossible, because of incomplete values.

As part of this new project configuration, we use a specific version of
Go, CUE and Node. These will be used as part of the release process, so
we need stability.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Ic22934d6939ae3e8bea482baeefaac5706143a78