all: add project configuration and use from CI
This is the beginning of a configuration that will be used to drive
various parts of the project. For now we import the project
configuration as part of the CI configuration, but it could/should
really be that the CI configuration is just a part of the project
configuration. That can wait until later.

Whilst here, drop the use of dummyDispatch for the trybot workflows. We
don't need that for now and it makes "testing" the GitHub workflows via
'cue export' impossible, because of incomplete values.

Also switch to assuming that the declared version of cmd/cue is
available in PATH, thus allowing us to "unify" on the use of 'cue' as
the go:generate command that is run, instead of having 'go run
cuelang.org/go/cmd/cue@$version' hard-coded all over the shop.

This then requires us to use the cue-lang/setup-cue action to install
CUE as part of CI.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Ic22934d6939ae3e8bea482baeefaac5706143a78