all: update extension properties ahead of publishing initial version
This brings the extension to be "valid" in the eyes of 'vsce publish'.
We also include the official CUE logo.

The name 'vscode-cue' is unique within the VSCode extension space.
Indeed in order to 'reserve' it ahead of landing this CL, I published an
early version.

The displayName of 'vscode-cue' is temporary. With the agreement of
another CUE-related extension in the marketplace, we plan to flip this
to 'CUE' at a later date.

The bump in vscode engine version is driven by our need/desire to
publish pre-release versions of vscode-cue.

With huge thanks to @tuxity for helping to get this started in
https://github.com/cue-lang/vscode-cue/pull/22.

This also includes the addition of a number of files to .vscodeignore in
order to not publish files unrelated to the extension. Many/all of
these, as identified in the comments, can be dropped when (in a later
CL) we move the extension to a subdirectory.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I75c17299bcee6d5d8959afd4d6f5324a5634d1f6
Reviewed-on: https://review.gerrithub.io/c/cue-lang/vscode-cue/+/1204087
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>