all: install `@vscode/vsce`
In preparation for publishing the first version of the extension.

As part of this, create a cue cmd workflow 'genManifest' that updates a
manifest.txt file that contains a list of the files included in the
extension. Correspondingly, add a step to the trybots that runs this
command.

Run 'vsce package' as a dry-run in the trybots.

We .vscodeignore *.cue files. For now we don't obviously need them; they
will act as the source for JSON etc.

Also .gitignore *.vsix files, the result of 'vsce package'.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I77eb757672030f3e4768c0b0cd55a86ff6d2e6f6