all: install `@vscode/vsce`
In preparation for publishing the first version of the extension.

As part of this, create a cue cmd workflow 'genManifest' that updates a
manifest.txt file that contains a list of the files included in the
extension.

Also define two npm script commands that package and publish. The
package step also runs the genManifest workflow.

Run 'npm run package' as a dry-run in the trybots.

We .vscodeignore *.cue files. For now we don't obviously need them; they
will act as the source for JSON etc.

Also .gitignore *.vsix files, the result of 'vsce package'.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I77eb757672030f3e4768c0b0cd55a86ff6d2e6f6
Reviewed-on: https://review.gerrithub.io/c/cue-lang/vscode-cue/+/1204088
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>