all: update LICENSE file to reflect history
@betawaffle kindly donated the original version of this plugin to the
CUE Project. It first moved to the cue-sh org, before finally moving to
cue-lang (the GitHub organisation that houses all the open source parts
of CUE).

Update the LICENSE file to reflect that changes since the "handover" are
copyright 'The CUE Authors', which is the group to which changes in the
open source CUE project are attributed.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I439c727742b3c3509fa2885c532476c86329dc46
Reviewed-on: https://review.gerrithub.io/c/cue-lang/vscode-cue/+/1204091
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>