internal/ci: switch to use new token-based auth for Central Registry
Align with the cue-lang/{cue,cuelang.org} repos in using the new
developer access tokens in combination with 'cue login --token' to
authenticate with the Central Registry.

In the process move away from the pattern we inherited from the
cue-lang/cue which seemed to imply that the auth persisted only as long
as the _registryReadOnlyAccessStep. This was not, and still is not, the
case, as a file was and is mutated.

Simplify instead to template this as a simple step.

Upgrade to CUE v0.11.0 in order to use the --token flag.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Iee6c7b80b80a1f10dac0a44705a7f1cbc9f7255d
Reviewed-on: https://review.gerrithub.io/c/cue-lang/vscode-cue/+/1204436
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>