internal/ci: update trybot workflow to also do releases
We plan to do releases as part of CI/CD for the VSCode extension. That
will be triggered by tagging the repository with a semantic version
string, e.g. 'v0.0.5'. This version must correspond to the declared
version of the extension in extension.cue.

This CL extends the trybot workflow to perform such a release on the
event of a 'v*' tag. It includes a CUE workflow command that performs
the check to ensure the tag on the repo matches the declared version
(noting that the declared version is only semver-like).

We set the VSCODE_PAT env var based on the cueckoo credentials establish
for the purpose of publishing.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I1cf4993017a67769f970bd7d727c4f23e69a6077
Reviewed-on: https://review.gerrithub.io/c/cue-lang/vscode-cue/+/1204445
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>