internal/ci: run 'npm install' instead of 'npm ci'
Use 'npm install' in CI, rather than 'npm ci', because the latter does
not appear to fail when, for example, the top-level 'version' field in
package.json is out of sync with package-lock.json. So whilst 'npm
install' is less efficient than 'npm ci' in a CI context, the two are
identical from a "which packages got installed" perspective, because we
always check that the repo is clean at the end of the workflow.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I808a4ea574165d7f251b20fe8429a6458c5cc39b