deps: update root package.json and package-lock.json
In a later CL, we switch to running 'npm install' as part of CI, in both
the root and extension directories (see comment in that CL for
reasoninig). That requires that output from 'npm install' be stable. A
lack of package name in the root package.json causes non-deterministic
output (depends on the directory in which the repository is checked
out). Fix that, and also ensure that package-lock.json is current
(something that was masked by 'npm ci' being used in the current stack,
and not failing when package-lock.json is not current).

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Iaf840715dfd8868957eff46644e5edc1785fff52