Fix GitHub issue/PR link
A #N reference in a commit message can only be a reference to an issue
or PR in this repository. To link to a CUE issue, it needs to be spelled
cue-lang/cue#N.

We keep a generic "issue/PR" and redirect to ${THIS_REPO}/issues/$N,
because that will correctly redirec to a PR if that number is a PR
rather than an issue. This will work OK for any repository whether or
not they have the issue tracker enabled.

That will work better longer term, particularly once we start templating
all of these project.config files across all repos. Otherwise we have to
distinguish "repos which have the issue tracker enabled" from those that
don't. When in practice we can configure both the same way.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I3327d40200e43c25f5c8d604ed59010036b3b32d