extension: set extension defaults in package.json
We current apply extension defaults in applyConfigDefaults, in code.
Because of https://github.com/Microsoft/vscode/issues/35451 this
basically doesn't work. Specifically, because
vscode.workspace.getConfiguration returns the zero value for a type in
case a configuration field is not set, there is no way to distinguish:

* A field not being set at any scope of configuration (global, workspace
  or folder);
* A default value equivalent to the zero value having been applied;
* The user setting a value equivalent to the zero value.

Put another way, for every field f defined in the configuration schema,
in the configuration object value returned, .has(f) will always return
true.

Right now I can't recall why I opted for code-level application of
defaults. But in short, it doesn't work because of this bug.

One possible way of working around this is to set a default value of
null in the configuration schema for each optional field. But even this
isn't perfect.

So instead we opt for setting default values in the configuration schema
found in package.json. But also reminding ourselves that we cannot ever
have logic in our extension that is predicated on a configuration field
not having been set.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Ia465154de9605bb5cc4adf67f1ed88e121a5e70c
Reviewed-on: https://review.gerrithub.io/c/cue-lang/vscode-cue/+/1206516
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>