extension: move from languageServerCommand to cueCommand
We currently have a config field for languageServerCommand, in order to
be able to specify a list of strings that represent the command to run
'cue lsp' (or equivalent).

However, this doesn't really fit with the fact that we asked the user to
install CUE (cmd/cue). The more common use case is someone wanting to
specify a specific path to cmd/cue.

Hence switch from languageServerCommand to cueCommand, and change to a
string type field. This means we unconditionally use the `lsp` command
as the first "argument" to cueCommand, with the existing
languageServerFlags config field then appended.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Iba3fa32cd7d76cc5232ada95b5ef52327552d53c