extension: check cueCommand path at config change time
Move the sanity check for the cueCommand value to the config change
handler. Moving this check here ensures that we can assert that after a
config change event, we can safely rely (in a happy path case) that the
value of cueCommand can be used, i.e. that it is a simple relative or
absolute path. The binary might still fail to resolve/exist but we
handle that at each call site.

In case we don't have a valid cueCommand value as a result of this
check, unconditionally stop the LSP. It will be started again when the
user provides a valid config value.

Also update the error handling approach in extensionConfigurationChange.
Because the caller of the handler does not handle a rejected promise, we
were swallowing errors. Switch instead to use a new helper method
showErrorMessage, which provides a void wrapper around
vscode.window.showErrorMessage.

Also provide and use a helper method absCueCommand which resolves the
cueCommand config value to an absolute path. This is used for now in
startCueLsp, but will also be used in later CLs.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Ifd6ded2b0006cf6ce8930dc75eb65276eef25da0
Reviewed-on: https://review.gerrithub.io/c/cue-lang/vscode-cue/+/1206539
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>