extension: check cueCommand path at config change time
Move the sanity check for the cueCommand value to the config change
handler. Moving this check here ensures that we can assert that after a
config change event, we can safely rely (in a happy path case) that the
value of cueCommand can be used, i.e. that it is a simple relative or
absolute path. The binary might still fail to resolve/exist but we
handle that at each call site.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Ifd6ded2b0006cf6ce8930dc75eb65276eef25da0