extension: use CUE as the source of truth for TypeScript
Currently we rely on keeping the TypeScript source and CUE in sync
manually. This is obviously brittle.

As a step in the right direction, generated a TypeScript file that
contains an exported version of the CUE configuration for the extension
(i.e. data, not schema). We do this via the new genTS command.

In order to be able to use the command IDs from TypeScript, we factor
out the list of commands required by the VSCode Extension schema into a
map, keyed by ID, and template the VSCode form from that map. We also
factor out the command prefix "CUE: " and modify the
Extension.registerCommand method to use the new config struct.

We also provide an Extension.commandID helper method to construct a
command ID from a command name. This is used again in a later CL.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Ifd6c76f101025da242edabb69049a7bbf7090a20
Reviewed-on: https://review.gerrithub.io/c/cue-lang/vscode-cue/+/1206549
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>