extension: dispose of lsp client state handler after stop
Now that we reflect the running (or otherwise) state of the LSP in the
status bar, we need to receive the event for the LSP stopping. Therefore
we need to dispose of the state change handler after the call to stop.

Caught in manual testing of vscode-cue.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I0f2dd8e4bd42b44aa69d60f9d4f2ec004e2b8ff1
Reviewed-on: https://review.gerrithub.io/c/cue-lang/vscode-cue/+/1206555
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>