extension: do not show error messages in startCueLsp
startCueLsp is called in response to configuration changes or when
manually attempting to start the LSP via a user command.

We cannot make assumptions about the calling context in startCueLsp as
to whether to show error messages or not. Instead, we should return a
rejected promise and let the caller decide what to do.

This CL implements that change.

Similarly, when handling a configuration state change, per the doc
comment for the handler, we need to show an error message to the user in
case of an error, rather than returning a rejected promise (because
the caller is not handling rejected promises, now confirmed via
experience).

This CL also implements that change.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Ib9cd8523ed30d8701a694e41c7edf207f3ef376f