deps: upgrade Go to v1.25.1
This allows us to access "go tool", through which we can enforce the
version of cmd/cue that's invoked.

The staticcheck version needed updating to support go1.25.1, so it's
added as a go-tool and invoked via a reuseable upstream step rather than
updating its direct go-run invocation version reference. A step from the
same upstream is used to abstract and normalise the go checks that are
performed.

Change-Id: I8ae8fd27f0e99acd29e1ee38e0099767785feee6
Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>
Reviewed-on: https://cue.gerrithub.io/c/cue-lang/vscode-cue/+/1222338
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>