extension: update description
This adds the key word "official", early in the extension's description,
in order to maximise its visibility when users use the in-editor search
UI to find a CUE-related extension and enhance the extension's
reputation.

It also tweaks the remainder of the description for the same
reputational reason, and to promote the extension's primary pair of
features.

Change-Id: I2ee4f49a3f7632727b29bc629cf62cd5ddf21a2f
Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>
Reviewed-on: https://cue.gerrithub.io/c/cue-lang/vscode-cue/+/1222340
Reviewed-by: Matthew Sackman <matthew@cue.works>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>