config: switch to using cue-lang/vscode-cue Approvers
We follow the pattern of cue-lang/cuelang.org by setting up a
"cue-lang/vscode-cue Approvers" group and permissioning that for +2,
Submit, adding of patch sets if not the owner. "CUE Approvers" is made a
member of this group. That way, we can add "custom" users outside of the
"CUE Approvers" group to the approvers list for this repo. The first
person being Jonathan.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Ia66923da13c4c39538ac0ac3d67a9eb71e72aa2b