extension: fix version checks in package-lock.json
CL 1222667 attempted to check that the value of the version field in
extension/extension.cue, serialised into extension/package.json, has
also been updated in two places in extension/package-lock.json.

Its implementation had a flaw because the "export" operation encoded in
extension/extension_tool.cue's genPackageJSON command suffers from
cue-lang/cue#4092. This issue strikes early in the go-generate process,
causing failure before extension/package.json is generated.

This change fixes that problem by moving the CUE constraints into its
own package, extension/vet, which is now explicitly targeted by a
separate, trailing, go-generate command in extension/gen.go. This
package placement is a little unfortunate, as it also requires all
cmd/cue invocations inside the extension working directory to be updated
to target the existing extension package, now that more than one package
is present in the directory.

The constraints' placement is forced, however, because they need to
embed the package-lock.json file, and therefore must exist either in the
file's directory or an ancestor - not a sibling or an ancestor's
sibling. Due to the ordering of "go generate ./..." (which invokes
commands in the same package order as "go list ./...", processing a
directory before its descendants), this ~forces the constraints to be
asserted by a cue-vet command in the extension directory. Whilst we
could separate the constraint location from the associated cue-vet
location, that would require the cue-vet command to include more brittle
".." directory traversals - which we choose to avoid at the cost of
slightly more verbose cmd/cue invocations.

Change-Id: I3a3235ef3e7b8555e54eaa8a874fb17de4dfa151
Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>