internal/ci: publish extension to OpenVSX
This adds a step to CI that publishes the extension to the OpenVSX
registry using the registry's ovsx command.

As per its documentation[0] the command is explicitly given the dynamic
filename of the packaged extension in order to skip a redundant
packaging step, and takes its credentials directly from the env.

Also: clarify two CI step names in passing.

[0] https://www.npmjs.com/package/ovsx#publish-extensions

Change-Id: I8835ea0fe1ea4a5c7c02120d4f7a45315cd81f33
Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>
Reviewed-on: https://cue.gerrithub.io/c/cue-lang/vscode-cue/+/1223012
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Reviewed-by: Paul Jolly <paul@myitcv.io>